fix(test): collapse rule-test 2-location split — HF-F (smoke tests now actually run)
ESLint rule tests previously had a 2-location split: smokeRule structural checks
lived at REPO-ROOT `tests/eslint-rules/<rule>.test.ts` (18 files), and the
RuleTester behavioural cases lived at `apps/admin/tests/eslint-rules/<rule>.test.ts`
(7 files). The `apps/admin/vitest.config.ts` is rooted at apps/admin and cannot
reach repo-root files, so the smoke checks shipped but NEVER RAN. The existence
ratchet (`check-eslint-rule-tests.ts`) was satisfied by file presence alone.

This is exactly the failure mode commit 0881b3ed flagged as the "HF-F follow-up"
when adding `no-secrets-in-client`:

  "the repo-root tests/eslint-rules/*.test.ts smoke files are existence-checked
   by the ratchet but NOT executed in CI (no root vitest config; the apps/admin
   vitest is rooted at apps/admin and cannot resolve them). The behavioral
   assertions live in apps/admin/tests/eslint-rules/ where they actually run.
   This is a pre-existing systemic gap affecting all 17 rules, logged for the
   HF-F follow-up."

This commit closes the gap. Approach: collapse the split. Every rule's test file
now lives in ONE location (`apps/admin/tests/eslint-rules/<rule>.test.ts`) and
carries BOTH the smoke check + (when present) the RuleTester behavioural cases.
existence-checked === actually-run.

Mechanics:
- Moved `_helpers.ts` from repo-root to `apps/admin/tests/eslint-rules/`.
- Moved 11 smoke-only files (the rules without RuleTester cases) from repo-root
  to `apps/admin/tests/eslint-rules/`, with the rule import path adjusted
  `"../../apps/admin/eslint-rules/<rule>.mjs"` → `"../../eslint-rules/<rule>.mjs"`.
- For the 7 rules that already had a RuleTester file at apps/admin: prepended the
  smokeRule `describe`/`it` block to the existing file (1 extra import + 5 lines)
  and deleted the repo-root smoke file. The 2 obsolete header-comment paragraphs
  ("Repo-root … carries the smokeRule structural check") were dropped.
- Updated `check-eslint-rule-tests.ts` TESTS_DIR to point at the new location,
  with HF-F context paragraph explaining why the split existed and how it was
  collapsed.
- Repo-root `tests/eslint-rules/` directory deleted entirely.

What HF-F surfaced (the closure receipt):
Re-running the 18 smoke tests as part of the apps/admin vitest revealed 5 LATENT
RULE DEFECTS that existed for months but never tripped CI because nothing was
running them:

  - no-ai-fanout-all, no-ai-forbidden-fields, no-hardcoded-greeting-in-composition,
    no-ops-import-from-api, no-orphan-instruction-fallback

All 5 are path-scoped rules whose `create()` early-returns `{}` (no AST visitors)
when the filename isn't on the guarded path. The smokeRule mock used a single
`getFilename: () => "/dev/null"` probe, which fails every isGuardedFile check, so
the rule returned 0 visitors and smokeRule's "at-least-one visitor" assertion
failed.

Fix: extended `smokeRule` to probe a list of plausible-guarded filenames
(`PROBE_FILENAMES`), taking the union of visitors returned. The assertion is now
"the rule returns visitors for AT LEAST ONE plausible guarded path", which
preserves the original intent (the rule CAN report) without baking rule-specific
paths into the helper. PROBE_FILENAMES has a comment header explaining "add a
fragment when a new rule lands whose guarded path is not yet represented".

The 5 defective tests all pass with the path-probing helper; no rule code changed.
All 87 tests across 18 files pass. The existence ratchet still holds 18/18.

KB updates:
- docs/kb/guard-registry.md — meta-ratchet block updated with the
  `check-eslint-rule-tests.ts` callout + HF-F context.
- docs/kb/generated/* regenerated for the file moves (new coupling edges).

## Verified by
- npx vitest run tests/eslint-rules/: 18 files, 87 tests, 0 failed.
- npx tsx scripts/capture/check-eslint-rule-tests.ts: 18/18, baseline 0.
- npm run kb:check: all 7 guards green (guard-links, rule-tests, webhook-sig,
  guard-tests, tsc-protected, knip-ratchet 161, fresh).
- npm run ratchet:check: tsc_errors 190 ✓, lint_errors 0 ✓. Pre-existing
  lint_warnings +2 drift unchanged from upstream 9c3f62d.
