+++ apps/admin/tests/eslint-rules/no-orphan-instruction-fallback.test.ts
@@ -1,5 +1,5 @@
 import { describe, it } from "vitest";
-import rule from "../../apps/admin/eslint-rules/no-orphan-instruction-fallback.mjs";
+import rule from "../../eslint-rules/no-orphan-instruction-fallback.mjs";
 import { smokeRule } from "./_helpers.js";
 
 describe("no-orphan-instruction-fallback", () => {
